:sparkles: feat : 주문 시, 사용할 적립금의 정상 범위가 벗어난 경우 예외 처리 기능 구현

+++ src/main/java/cart/dao/CartItemDao.java
@@ -32,7 +32,7 @@ public class CartItemDao {
             final long point = rs.getLong("point");
             Long productId = rs.getLong("product.id");
             String name = rs.getString("name");
-            int price = rs.getInt("price");
+            Long price = rs.getLong("price");
             String imageUrl = rs.getString("image_url");
             final double pointRatio = rs.getDouble("point_ratio");
             final boolean pointAvailable = rs.getBoolean("point_available");
@@ -75,7 +75,7 @@ public class CartItemDao {
             final long point = rs.getLong("point");
             Long productId = rs.getLong("id");
             String name = rs.getString("name");
-            int price = rs.getInt("price");
+            Long price = rs.getLong("price");
             String imageUrl = rs.getString("image_url");
             final double pointRatio = rs.getDouble("point_ratio");
             final boolean pointAvailable = rs.getBoolean("point_available");
